dodanie transakacji

+++ projekt_zespolowy/Controllers/TransactionsController.cs
@@ -12,7 +12,7 @@ using projekt_zespolowy.Models;
 
 namespace projekt_zespolowy.Controllers
 {
-    [Authorize]
+    [Authorize] // Blokuje dostęp dla niezalogowanych
     public class TransactionsController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -24,20 +24,22 @@ namespace projekt_zespolowy.Controllers
             _userManager = userManager;
         }
 
-        // GET: Transactions
-        public async Task<IActionResult> Index()
-        {
-            var currentUserId = GetCurrentUserId();
-
-            var userTransactions = _context.Transactions
-                    .Where(t => t.ApplicationUserId == currentUserId)
-                    .Include(t => t.Category);
-
-            return View(await userTransactions.ToListAsync());
-        }
+        // Metoda pomocnicza do pobierania ID aktualnego użytkownika
         private string GetCurrentUserId()
         {
             return _userManager.GetUserId(User);
+        }
+
+        // GET: Transactions
+        public async Task<IActionResult> Index()
+        {
+            var userId = GetCurrentUserId();
+            // Wyświetl transakcje tylko zalogowanego użytkownika
+            var transactions = _context.Transactions
+                .Include(t => t.Category)
+                .Where(t => t.ApplicationUserId == userId);
+
+            return View(await transactions.ToListAsync());
         }
 
         // GET: Transactions/Details/5
@@ -49,12 +51,19 @@ namespace projekt_zespolowy.Controllers
             }
 
             var transaction = await _context.Transactions
-                .Include(t => t.ApplicationUser)
                 .Include(t => t.Category)
+                .Include(t => t.ApplicationUser)
                 .FirstOrDefaultAsync(m => m.Id == id);
+
             if (transaction == null)
             {
                 return NotFound();
+            }
+
+            // Sprawdź czy transakcja należy do użytkownika
+            if (transaction.ApplicationUserId != GetCurrentUserId())
+            {
+                return Forbid();
             }
 
             return View(transaction);
@@ -63,21 +72,28 @@ namespace projekt_zespolowy.Controllers
         // GET: Transactions/Create
         public IActionResult Create()
         {
-            ViewData["ApplicationUserId"] = new SelectList(_context.Users, "Id", "Id");
-            ViewData["CategoryId"] = new SelectList(_context.Categories, "Id", "Name");
+            var userId = GetCurrentUserId();
+            // Pokaż tylko kategorie należące do użytkownika
+            ViewData["CategoryId"] = new SelectList(_context.Categories.Where(c => c.ApplicationUserId == userId), "Id", "Name");
             return View();
         }
 
         // POST: Transactions/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Description,Amount,Date,Type,CategoryId")] Transaction transaction)
+        public async Task<IActionResult> Create([Bind("Id,Description,Amount,Date,CategoryId")] Transaction transaction) // Usunięto "Type" z Bind
         {
-            transaction.ApplicationUserId = GetCurrentUserId();
+            var userId = GetCurrentUserId();
+            transaction.ApplicationUserId = userId;
 
+            // Ustawiamy domyślny typ transakcji na Wydatek (Expense), skoro użytkownik nie wybiera go w formularzu
+            transaction.Type = TransactionType.Expense;
+
+            // Usuwamy walidację dla powiązanych obiektów i typu
             ModelState.Remove("ApplicationUserId");
+            ModelState.Remove("ApplicationUser");
+            ModelState.Remove("Category");
+            ModelState.Remove("Type"); // Usuwamy walidację typu, bo ustawiamy go ręcznie
 
             if (ModelState.IsValid)
             {
@@ -86,10 +102,7 @@ namespace projekt_zespolowy.Controllers
                 return RedirectToAction(nameof(Index));
             }
 
-            ViewData["CategoryId"] = new SelectList(
-                _context.Categories.Where(c => c.ApplicationUserId == GetCurrentUserId()),
-                "Id", "Name", transaction.CategoryId);
-
+            ViewData["CategoryId"] = new SelectList(_context.Categories.Where(c => c.ApplicationUserId == userId), "Id", "Name", transaction.CategoryId);
             return View(transaction);
         }
 
@@ -106,22 +119,47 @@ namespace projekt_zespolowy.Controllers
             {
                 return NotFound();
             }
-            ViewData["ApplicationUserId"] = new SelectList(_context.Users, "Id", "Id", transaction.ApplicationUserId);
-            ViewData["CategoryId"] = new SelectList(_context.Categories, "Id", "Name", transaction.CategoryId);
+
+            // Sprawdź uprawnienia
+            if (transaction.ApplicationUserId != GetCurrentUserId())
+            {
+                return Forbid();
+            }
+
+            var userId = GetCurrentUserId();
+            ViewData["CategoryId"] = new SelectList(_context.Categories.Where(c => c.ApplicationUserId == userId), "Id", "Name", transaction.CategoryId);
             return View(transaction);
         }
 
         // POST: Transactions/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Description,Amount,Date,Type,CategoryId,ApplicationUserId")] Transaction transaction)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Description,Amount,Date,Type,CategoryId")] Transaction transaction)
         {
             if (id != transaction.Id)
             {
                 return NotFound();
             }
+
+            // Pobierz oryginał z bazy, żeby sprawdzić właściciela (bez śledzenia zmian na razie)
+            var existingTransaction = await _context.Transactions.AsNoTracking().FirstOrDefaultAsync(t => t.Id == id);
+
+            if (existingTransaction == null)
+            {
+                return NotFound();
+            }
+
+            if (existingTransaction.ApplicationUserId != GetCurrentUserId())
+            {
+                return Forbid();
+            }
+
+            // Upewnij się, że ID użytkownika się nie zmienia
+            transaction.ApplicationUserId = GetCurrentUserId();
+
+            ModelState.Remove("ApplicationUserId");
+            ModelState.Remove("ApplicationUser");
+            ModelState.Remove("Category");
 
             if (ModelState.IsValid)
             {
@@ -143,8 +181,9 @@ namespace projekt_zespolowy.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["ApplicationUserId"] = new SelectList(_context.Users, "Id", "Id", transaction.ApplicationUserId);
-            ViewData["CategoryId"] = new SelectList(_context.Categories, "Id", "Name", transaction.CategoryId);
+
+            var userId = GetCurrentUserId();
+            ViewData["CategoryId"] = new SelectList(_context.Categories.Where(c => c.ApplicationUserId == userId), "Id", "Name", transaction.CategoryId);
             return View(transaction);
         }
 
@@ -157,12 +196,18 @@ namespace projekt_zespolowy.Controllers
             }
 
             var transaction = await _context.Transactions
-                .Include(t => t.ApplicationUser)
                 .Include(t => t.Category)
+                .Include(t => t.ApplicationUser)
                 .FirstOrDefaultAsync(m => m.Id == id);
+
             if (transaction == null)
             {
                 return NotFound();
+            }
+
+            if (transaction.ApplicationUserId != GetCurrentUserId())
+            {
+                return Forbid();
             }
 
             return View(transaction);
@@ -174,18 +219,26 @@ namespace projekt_zespolowy.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var transaction = await _context.Transactions.FindAsync(id);
+
             if (transaction != null)
             {
+                if (transaction.ApplicationUserId != GetCurrentUserId())
+                {
+                    return Forbid();
+                }
+
                 _context.Transactions.Remove(transaction);
+                await _context.SaveChangesAsync();
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool TransactionExists(int id)
         {
-            return _context.Transactions.Any(e => e.Id == id);
+            // Sprawdza istnienie transakcji należącej do tego użytkownika
+            var userId = GetCurrentUserId();
+            return _context.Transactions.Any(e => e.Id == id && e.ApplicationUserId == userId);
         }
     }
 }
